+++ work-dir/python/Zadaci_nastavak/Tacka4/.idea/.gitignore
@@ -1,0 +1,3 @@
+# Default ignored files
+/shelf/
+/workspace.xml
